requirements.txt
django
gunicorn
djangorestframework
django-cors-headers
python-dotenv
psycopg2
